requirements.txt
flask
flask-cors
openai
pdfplumber
python-docx
pandas
openpyxl
pytesseract
pillow
faiss-cpu
numpy
python-dotenv
gunicorn  # Recommended for Render
